Headless and GLFW Windows (#491)
* window interface

* headless window

* GLFW window

* add VKB_TEST_GLFW

* fix glfw errors

* add samples_launcher window improvements

* STD Filesystem VFS (#495)

* std filesystem

more tests and fixes

recursive tests

more tests

remove order specific checks from tests

review comments

* fix duplicated check

* review changes

* GLFW window

* add VKB_TEST_GLFW

* directly create a surface

* Direct to Display

+++ components/events/include/components/events/input_manager.hpp
@@ -29,8 +29,8 @@ namespace events
 {
 struct CursorPosition
 {
-	uint32_t x;
-	uint32_t y;
+	float x;
+	float y;
 };
 
 struct Touch
